fix(sandbox): revert example-nextjs, simplify to single example page
- Revert example-nextjs to node_modules paths (it intentionally
  represents an external consumer)
- Consolidate buttons/form/typography pages into a single example page
  that serves as a scaffold for designers
- Add trailing slashes to Sidebar hrefs and document the convention
  for static exports
