+++ apps/example-nextjs/postcss.config.js
@@ -7,7 +7,6 @@ module.exports = {
       include: [
         'src/**/*.{js,jsx,ts,tsx}',
         'node_modules/@xds/core/**/*.{ts,tsx}',
-        '../../packages/theme/src/**/*.{ts,tsx}',
       ],
       babelConfig: {
         babelrc: false,
@@ -16,8 +15,6 @@ module.exports = {
         },
         presets: [
           ['@babel/preset-react', {runtime: 'automatic'}],
-          // Must come after preset-react (runs first due to reverse order)
-          // to strip TypeScript type assertions before StyleX evaluates them.
           '@babel/preset-typescript',
         ],
         plugins: babelConfig.plugins,
